backend/requirements.txt
fastapi 
uvicorn[standard]
openai
tzlocal
python-multipart
requests
PyPDF2
python-docx
openpyxl
striprtf
odfpy
aiohttp
duckduckgo_search
python-pptx
tavily-python
pywin32; sys_platform == 'win32'
bs4
langchain_openai
langchain_ollama
langchain_community
langchain-google-community
langchain-exa
faiss-cpu
mcp
nest_asyncio
shortuuid
docstring_parser
python-a2a[all]>= 0.5.6
ollama
appdirs
aiosqlite
rank_bm25
fastapi_mcp
mem0ai
e2b-code-interpreter
qq-botpy
scipy
edge-tts
Brotli
wikipedia-api
arxiv
ddgs
claude-agent-sdk
aiofiles
sherpa_onnx
soundfile
nvidia-ml-py
pip-licenses 
lark_oapi
pydub
imageio_ffmpeg
discord.py
coqui-tts>=0.22.0
diffusers>=0.25.0
accelerate>=0.25.0